bug_hunter/requirements.txt
# Bug Hunter Tool Requirements

# Core dependencies
requests>=2.31.0
pyyaml>=6.0
dnspython>=2.4.0

# AI/ML dependencies
google-generativeai>=0.8.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Networking
python-nmap>=0.7.1
scapy>=2.5.0

# Reporting
jinja2>=3.1.0
markdown>=3.5.0
weasyprint>=60.0

# Utilities
colorama>=0.4.6
tqdm>=4.66.0
click>=8.1.0

# Security
cryptography>=41.0.0
pyopenssl>=23.0.0

# Optional dependencies for enhanced features
# Uncomment if needed:
# selenium>=4.15.0  # For browser automation
# playwright>=1.40.0  # Alternative browser automation
# shodan>=1.30.0  # For Shodan integration
# censys>=2.2.0  # For Censys integration

